CI: generate and publish the documentation
The documentation is generated invoking the Doxyfile and is than
moved to /srv/code_documentation on the gitlab server replacing the
previous documentation of the same project and same branch.
Folder are structured so that is possible to publish documentation
(or other files) from the pipeline execution of different projects
and different branches.
The documentation is reachable via https://documentation.skywarder.eu/