[Registry-Frontend] Safe and unsafe interface and type data structures update
Now the interface for the front-end has type safe and type unsafe methods. For now all will be as unsafe in the implementation but the interface does contain such type-safe method signatures.
Added some type data structures. Now no more MACROS for the data structures but inheritance and templates are used instead.