[Registry-Frontend] RegistryStructure.h added w/ wrappers for union and .cpp WIP
Added a specific RegistryStructure.h file for the structures specific to the configuration, the type union, the configuration enum and the wrappers datastructure for the datatypes.
Implementation on the .cpp file reguarding the get/set both safe and unsafe.
Renaming to Unsafe methods and no more adding the safe specification to the signature of the type safe methods.
Unsafe method have their specific getters/setters inside the class for the registry front-end to get/set the union type from it.
Also added to the cmake the RegistryFontEnd.cpp path.