[Registry-Frontend] EntryStructUnion privates methods, Load method, Removed methods
EntryStructUnion: Now the privates method to get/set the union is in the EntryStructUnion. Now using a type enumerator field for checking the correct type even for unsafe methods.

Load method: Now there will be a load method to get the configuration from the backend and check corruption
Removed method: Removed set/get methods with struct template. Now using the number and the type only for the unsafe interface.