[Registry-Frontend] Concurrency management for non-blocking write to Backend
Now the registry frontend has writeBuffers structures with their own mutex and vector. While one is used for writing the current configuration, the other is used as buffer to be written to backend. By doing so, with a variable while one copy is blocked for the backend write, the other is at disposition of the backend to write it on memory, without having blocking behaviors.