[Registry-Frontend] Added the registry-middleware + save at each set + updateSerializedData
Middleware: A new .h a .cpp for the RegistryMiddleware and RegistryMiddlewareFlash classes. Those are active object that the frontend can call to have thread safe writes to the backend without blocking while erasing and writing to the underlying hw.
Save at each set: Now the RegistryFrontend saves at each new set, calling the middleware to write the new data.
UpdateSerializedData: Now instead of getting the reference to the vector we do not expose that and have just an update method.