[USART] Call Miosix IRQ handlers instead of accessing driver implementation details
In order to reduce the number of patches we make to the Miosix kernel, we're moving to calling Miosix USART IRQ handlers instead of accessing static port objects directly, which would require marking them for external linkage.
With this commit we are dropping support for USART ports higher than 3 when using the Miosix USART driver. For higher port counts, the boardcore driver must be used.