[Registry-Frontend] Fix: Inhibit saving during armed state + private vector update
Fix: Now the saving command can be triggered only if not in armed state. This is to prevent vector allocations during armed phase.
Private update: The update of the serialized vector has been privatized, since it is used before the saving and it must be inhibited in case of armed phase.