[Registry-Frontend] RegistrySerializer class with header struct + re-organize types
RegistrySerializer: Introduced a RegistrySerializer class for the serialization and de-serialization of the vector, since semantically is more correct to have an appropriate class while the RegistryFrontend does not implements such serialization mechanisms.
RegistryHeader: Now the header is not handmade but is mapped onto an appropriate struct.
Re-organize types: To avoid RegistrySerializer to depend on the RegistryFrontend.h file for the types, they where moved into the already existing TypeStructures.h file.
WIP note: This still is a WIP since now it is necessary to use such RegistrySerializer class from the RegistryFrontend one.