[HILSensor] Implemented the HILSensor class and the hillificator function
The HILSensor class, templated on a real sensor of type T, overrides the sampleImpl method to use the passed mapping function to generate the last sample from simulator data. The init and selfTest methods are overridden so that the real methods are called only if the "hardware is enabled". Also, a getRealLastSample method is implemented so that, in case that the "hardware is enabled", we can get also the corresponding real sample.

The hillificator function is an helper function that, once passed the sensor pointer, the flag to enable the hardware and the mapping function, returns the corresponding HILSensor. The returned sensor behaves just like the real sensor, but the sample that it returns comes from the updateData function (which usually maps the simulator data to the data returned by the sensor).