[TaskScheduler] Rework SKIP policy to only skip past executions
Previously, when a task with SKIP policy missed one or more executions, the scheduler would skip missed executions, but it would also skip running the task altogether if the current timestamp was not aligned to the task period. This caused the task to never execute if the system load was high enough that the timestamp was never aligned to the task period. This issue was aggravated by the the move to nanosecond periods, as the timestamps would never match.
Now the scheduler skips missed executions, but it runs the task once, despite the current timestamp not being aligned to the task period.