[Follower] Add follower algorithm from obsw
[Follower] Add minimize rotationa as suggested in previous obsw pull req.

See [this](https://git.skywarder.eu/avn/swd/obsw/-/merge_requests/34)
for more context

[logger] Updated logger

[ARP Follower] Yaw gain now 0.1

The yaw gain gain now is no more 1 but fixed to 0.1

[ARP] Added timestamp to AntennaAngles for log

Added timestamp to the AntennaAngles logged. Changed accordingly the constructors.

[Fix] Follower algorithm limitation issue due to proportional on yaw

The 0.1* product to have a proportional on the yaw was creating issues with the angle limiting function.

[ARP] Added nrPropagations for log in AntennaAngle now logged by SMA

Added nrPropagations to AntennaAngles
Moved the log of it from the Follower to SMA

[ARP] Follower initRocketCoordinates now is NASOrigin

Changed semantic of initialRocketCoordinates to NASOrigin since now it is set always as we have a NASOrigin.

[ARP] WIP fix for online initial coordinates support

Add support for initial coordinates online setting

[ARP] Follower Mutexes for antenna and NAS origin's coordinates

Now getter states to protect such variables with mutexes