[Logger] New logger with socrate
We have decided to use socrate instead of tscpp for logging because
it allows for more flexibility.

When a new type is logged the system also logs a mapping string that
encodes the structure of that type. This string is then used in the
deserializer to get the structure of the type and deserialize it
correctly.