[Logger] New logger with socrate
We have decided to use socrate instead of tscpp for logging because
it allows for more flexibility.

The new system creates both a log and a mapping file which contains
a description of all of the types that have been logged. Both of
these files are used in the Deserializer class to deserialize the logs
without the need to specify the structure of each type in the code.