[SignaledDeadlineTask] Introduce a signaled deadline task class
This is a task that executes a user-defined function at specific time
points, or when signaled.
This class is designed to be used in situations where a task needs to be
executed within specific deadlines, irregularly. The task will wake up
at the specified time points and execute the user-defined function. The
task can also be signaled to wake up and run immediately, regardless of
the time points.

Co-authored-by: Pietro Bortolus <pietro.bortolus@skywarder.eu>