[Utils] Introduce Google's bfloat16 type
This is a 16-bit floating point type, resulting from the truncation of
the last 16 least significant bits of the mantissa.
This may come in handy for reducing bandwidth usage of radio telemetry,
as this effectively cuts in half the size requirements for sending
floating point values, which are the most used types of data for us.