[ND0xxx] fixed error when sending config to ND0xxx sensors
This fix applies to all of the ND pressure sensors (ND015A, ND015D and ND030D)
The data sent to the sensor to configure it was in the wrong order, the datasheet
demanded the MSB to be sent first, this was fixed by putting the odr in the
sensorConfig bitfield as the first field.

In addition to changing the order of the sensorConfig bitfield bit_cast is
used instead of static_cast as the ladder is implementation defined and
could technically yield the wrong result, furthermore using a bit_cast is
semantically more correct.

Additionally, methods to set, update, and get an offset, that is added to
the computed pressure, were added.