[USART] Implemented proper error flag handling
Previously, in the IRQHandleInterrupt() function, the code only cleared the IDLE flag from the ICR (Interrupt Clear Register) which meant that if another flag, like FE (Framing Error), was triggered it would never be cleared. This meant that after detecting an error, which would normally only discard the byte that is being read, the driver would keep ignoring all incoming bytes until the ISR (Interrupt Status Register) got cleared by a reset.

This issue was solved by writing in the ICR (Interrupt Clear Register) to clear all error flags in the IRQHandleInterrupt() function.