[Ereg] Implemented electronic pressure regulator algorithm
The goal of the algorithm is to control a valve's servo to regulate the
downstream pressure based on an upstream and a downstream pressure reading.
The algorithm is based on a PID controller, the output of the PID is then
converted to a servo command value.

The algorithm allows to set the initial integral contribution for the PID.