[algorithms] Added an algorithm for an electronic regulator
This algorithm takes a pressure input downstream from a valve and uses a simple PID to control the
valve's servo to regulate the downstream pressure.
It is possible to change the reference point and the PID constants while the algorithm is running,
doing so resets the internal state of the algorithm.